requirements_dev.txt
bumpversion==0.5.3
wheel==0.29.0
django-model-utils>=2.0
ipdb==0.10.3
